Move libraryHandler to collection/drive
