+++ src/internal/m365/collection/drive/library_handler_test.go
@@ -1,4 +1,4 @@
-package site
+package drive
 
 import (
 	"testing"
