+++ src/internal/m365/collection/site/restore.go
@@ -41,7 +41,7 @@ func ConsumeRestoreCollections(
 	ctr *count.Bus,
 ) (*support.ControllerOperationStatus, error) {
 	var (
-		lrh            = libraryRestoreHandler{ac}
+		lrh            = drive.NewLibraryRestoreHandler(ac)
 		restoreMetrics support.CollectionMetrics
 		caches         = drive.NewRestoreCaches(backupDriveIDNames)
 		el             = errs.Local()
