+++ src/internal/operations/backup_test.go
@@ -20,8 +20,8 @@ import (
 	"github.com/alcionai/corso/src/internal/kopia"
 	"github.com/alcionai/corso/src/internal/m365/graph"
 	"github.com/alcionai/corso/src/internal/m365/mock"
-	odConsts "github.com/alcionai/corso/src/internal/m365/onedrive/consts"
-	odMock "github.com/alcionai/corso/src/internal/m365/onedrive/mock"
+	odMock "github.com/alcionai/corso/src/internal/m365/service/onedrive/mock"
+	odConsts "github.com/alcionai/corso/src/internal/m365/service/onedrive/consts"
 	"github.com/alcionai/corso/src/internal/m365/support"
 	"github.com/alcionai/corso/src/internal/model"
 	"github.com/alcionai/corso/src/internal/operations/inject"
